Initial commit
Initial commit of lens library for Racket. Still needs work to fix
fundamental issues: In particular, locations that either don't exist or
are the wrong type, which I am now seeing as two sides of the same coin.
Also, struct-lens doesn't work. Prisms need more investigation, and more
lenses need to be added, and unit tests need to be written, etc. etc.
etc. This is just a start.